extract map positions picker to shared directory
