+++ client/src/shared/MapPositions/index.js
@@ -1,0 +1,2 @@
+export { getMapPositionName, getMapPositionColor } from './mapPosition'
+export { MapPositions } from './MapPositions'
