+++ client/src/shared/index.js
@@ -1,3 +1,3 @@
 export { PointsWithDelta } from './PointsWithDelta'
-export { getMapPositionName, getMapPositionColor } from './mapPosition'
+export { getMapPositionName, getMapPositionColor } from './MapPositions'
 export { ColorBox } from './ColorBox'
